// next-js/src/lib/nextpress/types/acf/components/FieldProps.d.ts
type FieldProps<LayoutT> =
    LayoutT extends { sub_fields: readonly any[] }
        ? ResolveFields<LayoutT['sub_fields']> & { post: WPPost }
    : never;

type ResolveFields<Fields extends readonly any[]> = {
    [F in Fields[number] as F['name']]: MapFieldType<F>
};

type ResolveLayouts<Layouts extends readonly any[]> = {
    [L in Layouts[number] as L['name']]: L extends { name: infer Name, sub_fields: readonly any[] }
        ? { layout: Name } & ResolveFields<L['sub_fields']>
    : never
}[Layouts[number]['name']];


type MapFieldType<Field> =
    Field extends { type: 'color_picker' }
        ? string | null
    : Field extends { type: 'date_picker' }
        ? string | null
    : Field extends { type: 'date_time_picker' }
        ? string | null
    : Field extends { type: 'google_map' }
        ? ACFGoogleMapsObject | null
    : Field extends { type: 'icon_picker' }
        ? Field extends { return_format: 'string' }
            ? string | null
        : Field extends { return_format: 'array' }
            ? ACFIconObject | null
        : ACFIconObject | null
    : Field extends { type: 'time_picker' }
        ? string | null
    : Field extends { type: 'email' }
        ? string | null
    : Field extends { type: 'number' }
        ? number | null
    : Field extends { type: 'password' }
        ? string | null
    : Field extends { type: 'range' }
        ? number | null
    : Field extends { type: 'text' }
        ? string | null
    : Field extends { type: 'textarea' }
        ? string | null
    : Field extends { type: 'button_group' }
        ? Field extends { return_format: 'array' }
            ? ACFChoiceObject | null
        : string | null
    : Field extends { type: 'checkbox' }
        ? Field extends { return_format: 'array' }
            ? ACFChoiceObject[]
        : string[]
    : Field extends { type: 'radio' }
        ? Field extends { return_format: 'array' }
            ? ACFChoiceObject | null
        : string | null
    : Field extends { type: 'select' }
        ? Field extends { multiple: 1 }
            ? Field extends { return_format: 'array' }
                ? ACFChoiceObject[]
            : string[]
        : Field extends { return_format: 'array' }
            ? ACFChoiceObject | null
        : string | null
    : Field extends { type: 'true_false' }
        ? boolean
    : Field extends { type: 'file' }
        ? Field extends { return_format: 'id' }
            ? number | null
        : Field extends { return_format: 'url' }
            ? string | null
        : never
    : Field extends { type: 'gallery' }
        ? Field extends { return_format: 'id' }
            ? number[]
        : Field extends { return_format: 'url' }
            ? string[]
        : never
    : Field extends { type: 'image' }
        ? Field extends { return_format: 'id' }
            ? number | null
        : Field extends { return_format: 'url' }
            ? string | null
        : never
    : Field extends { type: 'oembed' }
        ? string | null
    : Field extends { type: 'wysiwyg' }
        ? string | null
    : Field extends { type: 'link' }
        ? Field extends { return_format: 'url' }
            ? string | null
        : Field extends { return_format: 'array' }
            ? ACFLinkObject | null
        : ACFLinkObject
    : Field extends { type: 'page_link' }
        ? Field extends { multiple: 1 }
            ? string[]
        : string | null
    : Field extends { type: 'post_object' }
        ? Field extends { multiple: 1 }
            ? Field extends { return_format: 'object' }
                ? WPPost[]
            : number[]
        : Field extends { return_format: 'object' }
            ? WPPost | null
        : number | null
    : Field extends { type: 'relationship' }
        ? Field extends { return_format: 'object' }
            ? WPPost[] | null
        : number[] | null
    : Field extends { type: 'taxonomy' }
        ? Field extends { multiple: 1 }
            ? Field extends { return_format: 'object' }
                ? WPTerm[]
            : number[]
        : Field extends { return_format: 'object' }
            ? WPTerm | null
        : number | null
    : Field extends { type: 'user' }
        ? Field extends { multiple: 1 }
            ? Field extends { return_format: 'object' }
                ? WPUser[]
            : number[]
        : Field extends { return_format: 'object' }
            ? WPUser | null
        : number | null
    : Field extends { type: 'group'; sub_fields: readonly any[] }
        ? ResolveFields<Field['sub_fields']> | null
    : Field extends { type: 'repeater'; sub_fields: readonly any[] }
        ? ResolveFields<Field['sub_fields']>[] | null
    : Field extends { type: 'flexible_content'; layouts: readonly any[] }
        ? ResolveLayouts<Field['layouts']>[] | null
    : Field extends { sub_fields: readonly any[] }
        ? ResolveFields<Field['sub_fields']>
    : never;